refactor: implement config::value::Value, which can handle secrets more securely, and introduce extesionability for future. eg. value = "${file(path/to/file)}" to read value from file
