+++ core/src/config/vcs.rs
@@ -1,0 +1,33 @@
+use std::fmt;
+
+use serde::{Deserialize, Serialize};
+
+use crate::config;
+
+#[derive(Serialize, Deserialize)]
+#[serde(tag = "type")]
+pub enum Account {
+    #[serde(rename = "github")]
+    Github {
+        email: config::Value,
+        account: config::Value,
+        key: config::Value
+    },
+    #[serde(rename = "gitlab")]
+    Gitlab {
+        email: config::Value,
+        account: config::Value,
+        key: config::Value
+    },
+    #[serde(rename = "module")]
+    Module(config::module::ConfigFor<crate::vcs::Manifest>)
+}
+impl fmt::Display for Account {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Github{..} => write!(f, "github"),
+            Self::Gitlab{..} => write!(f, "gitlab"),
+            Self::Module{..} => write!(f, format!("module {}", self.uses)),
+        }
+    }    
+}
