+++ docs/new_config.toml
@@ -15,11 +15,22 @@ lab = { patterns = ["lab"] }
 prod = { tag = true, patterns = ["[0-9]*"] }
 
 # ------------
+# secret settings
+# ------------
+[secrets]
+type = "dotenv"
+keys = [
+    "SUNTOMI_VCS_ACCOUNT",
+    "SUNTOMI_VCS_ACCOUNT_EMAIL",
+    "SUNTOMI_VCS_ACCOUNT_KEY"
+]
+
+# ------------
 # vcs settings
 # ------------
 [vcs]
 # version control system account settings
-type = "Github"
+type = "github"
 email = "${SUNTOMI_VCS_ACCOUNT_EMAIL}"
 account = "${SUNTOMI_VCS_ACCOUNT}"
 key = "${SUNTOMI_VCS_ACCOUNT_KEY}"
@@ -31,7 +42,7 @@ key = "${SUNTOMI_VCS_ACCOUNT_KEY}"
 [ci.default]
 # ci account setting, mutiple ci service can be used by creating mutiple [ci.accounts.$name] sections.
 # [ci.default] should always be set.
-type = "GhAction"
+type = "ghaction"
 account = "${SUNTOMI_VCS_ACCOUNT}"
 key = "${SUNTOMI_VCS_ACCOUNT_KEY}"
 
@@ -57,7 +68,7 @@ repository = {
     }
 }
 slack = {
-    use = "suntomi/slack@0.2.0",
+    uses = "suntomi/slack@0.2.0",
     with = {
         webhook = "https://hooks.slack.com/services/T0J5QJQ0P/B0J5QJQ0P/X0J5QJQ0P",
         channel = "#suntomi",
@@ -67,10 +78,10 @@ slack = {
 }
 discord = {
     type = "Custom",
-    use = "suntomi/discord@0.3.0",
+    uses = "suntomi/discord@0.3.0",
     with = {
-
-    }    
+        key = "${SUNTOMI_DISCORD_KEY}",
+    }
 }
 
 # ------------
@@ -82,7 +93,7 @@ discord = {
 on = { workflows = ["deploy", "integrate"], changed = ["tools/docker/Dockerfile.base"] }
 # you can set local_fallback container image. if its set, command will be executed on container of `local_fallback.image`,
 # with using shell `local_fallback.shell`, if local machine's OS does not matched runner.os.
-runner = { type = "Machine", os = "Linux", local_fallback = { image = "docker:20.10", shell = "sh" } }
+runner = { type = "machine", os = "Linux", local_fallback = { image = "docker:20.10", shell = "sh" } }
 command = """
 docker build -t ghcr.io/suntomi/deplo:base -f tools/docker/Dockerfile.base tools/docker
 if [ "${DEPLO_CI_WORKFLOW_TYPE}" = "deploy" ]; then
@@ -94,7 +105,7 @@ fi
 [jobs.builder]
 # if workflows is opmitted, use ["deploy", "integrate"] by default.
 on = { changed = ["tools/docker/Dockerfile.builder", "Cargo.*"] }
-runner = { type = "Machine", os = "Linux", local_fallback = { image = "docker:20.10", shell = "sh" } }
+runner = { type = "machine", os = "Linux", local_fallback = { image = "docker:20.10", shell = "sh" } }
 command = """
 sh tools/scripts/copy_manifests.sh
 docker build -t ghcr.io/suntomi/deplo:builder -f tools/docker/Dockerfile.builder tools/docker
@@ -122,7 +133,7 @@ on = [
     { workflows = ["repository"], events = ["manual"] },
     { workflows = ["slack"], when = { message = { matches = [".*test\sit.*"] } } }
 ]
-runner = { type = "Machine", os = "Linux", local_fallback = { image = "ghcr.io/suntomi/deplo:builder" } }
+runner = { type = "machine", os = "Linux", local_fallback = { image = "ghcr.io/suntomi/deplo:builder" } }
 command = "cargo test"
 tasks = { test = "cargo test", home = "echo ${HOME}" }
 # cache settings. file paths that is listed in paths will be preserved with tha key `keys[0]`.
@@ -142,21 +153,21 @@ caches = {
 on = { workflows = ["deploy"], changed = ["Deplo.toml"] }
 # local_fallback can be path to docker file. if you specify docker file path, deplo automatically build corresponding image
 # and use it as local fallback container image.
-runner = { type = "Machine", os = "Linux", local_fallback = { path = "tools/docker/Dockerfile.vmfb", shell = "sh" } }
+runner = { type = "machine", os = "linux", local_fallback = { path = "tools/docker/Dockerfile.vmfb", shell = "sh" } }
 command = "deplo init"
 # commits settings. if path of files that changed after the job finish running, matched patterns, 
 # deplo push or made pull request such a changes to repoisitory, after all jobs finish running.
 commits = [
-    { patterns = [".github/*"], for_targets = ["nightly"], push_opts = { type = "Push" } },
-    { patterns = [".github/*"], for_targets = ["prod"], push_opts = { type = "PullRequest" } },
-    { patterns = ["docs/*"], for_targets = ["lab"], push_opts = { type = "Push" } }
+    { patterns = [".github/*"], targets = ["nightly"], method = { type = "push" } },
+    { patterns = [".github/*"], targets = ["prod"], method = { type = "pull_request" } },
+    { patterns = ["docs/*"], targets = ["lab"], method = { type = "push" } }
 ]
 
 [jobs.product]
 on = { workflows = ["deploy"], changed = ["*/src/*", "Cargo.*"] }
 # job dependencies. deploy.product only starts to run after deploy.builder finished.
 depends = ["builder"]
-runner = { type = "Machine", os = "Linux", local_fallback = { path = "tools/docker/Dockerfile.vmfb", shell = "sh" } }
+runner = { type = "machine", os = "linux", local_fallback = { path = "tools/docker/Dockerfile.vmfb", shell = "sh" } }
 command = """
 set -e
 mkdir -p tools/docker/bin
@@ -189,7 +200,7 @@ on = { workflows = ["deploy"], targets = ["prod"], changed = ["*/src/*", "Cargo.
 # prohibit to run from cli
 options = { only_from_ci = true }
 depends = ["product"]
-runner = { type = "Machine", os = "Linux", local_fallback = { image = "docker:20.10", shell = "sh" } }
+runner = { os = "Linux", local_fallback = { image = "docker:20.10", shell = "sh" } }
 command = """
 . tools/scripts/setup_release_env.sh
 echo ${SUNTOMI_VCS_ACCOUNT_KEY} | docker login ghcr.io -u ${SUNTOMI_VCS_ACCOUNT} --password-stdin
@@ -201,7 +212,7 @@ docker push ghcr.io/suntomi/deplo:latest
 [jobs.mac]
 on = { workflows = ["deploy"], changed = ["*/src/*", "Cargo.*"] }
 depends = ["product"]
-runner = { type = "Machine", os = "MacOS" }
+runner = { type = "machine", os = "macos" }
 command = """
 set -e
 . tools/scripts/setup_release_env.sh
@@ -223,7 +234,7 @@ caches = {
 [jobs.win]
 on = { workflows = ["deploy"], changed = [".*/src/.*", 'Cargo\.*'], diff_matcher = "regex" }
 depends = ["product"]
-runner = { type = "Machine", os = "Windows" }
+runner = { os = "windows" }
 command = """
 set -e
 . tools/scripts/setup_release_env.sh
@@ -256,8 +267,10 @@ config::ci::Account
 config::vcs::Config
 config::vcs::Account
 
-bootstrap flow 
+secret::Secret
+secret::dotenv::Dotenv
 
+bootstrap flow 
 1. create config::runtime::Config
   - and apply 1st priority config like log level
 2. load Deplo.toml
@@ -267,3 +280,11 @@ bootstrap flow
 4. allow secret module to load and inject secret into config::Config
 5. allow module to modify config
 6. invoke actual command 
+
+
+ci kick flow 
+1. create dependency graph of jobs
+2. determine workflow type and parameters (deploy, integrate, slack, discord, etc...)
+3. determine target, if exists (dev, stage, prod, etc...)
+4. filter jobs by workflow, its parameters, and target and mark execute or not execute for each node of graph
+5. invoke jobs according to their dependencies
